pyproject.toml
[project]
name = "envibe"
version = "0.dev0"
authors = [{ name = "luwqz1", email = "howluwqz1@gmail.com" }]
description = "Modern tiny env parser."
readme = "README.md"
requires-python = ">=3.13"
keywords = [
  "env parser",
  "type-hinted",
  "environ",
  "tiny",
  "fp",
  "functional programming",
]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
  "Typing :: Typed",
]
dependencies = [
    "felis-catus>=0.dev5",
]

[dependency-groups]
dev = [
  "ruff>=0.11.2,<0.12",
  "basedpyright>=1.28.3,<2",
  "sort-all>=1.3.0,<2",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 160
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
  "I",
  "D",
  "N",
  "PLR",
  "Q",
  "COM",
  "TC",
  "YTT",
  "SIM"
]
ignore = [
  "COM812",
  "N805",
  "N818",
  "TC001",
  "TC002",
  "TC003",
  "TC004",
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D107",
  "D202",
  "D203",
  "D205",
  "D209",
  "D211",
  "D213",
  "D400",
  "D401",
  "D404",
  "D415",
  "PLR2004",
  "PLR0911",
  "PLR0912",
  "PLR0913"
]
fixable = ["ALL"]
exclude = [".venv", ".idea"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.pyright]
exclude = [
  ".venv",
  ".idea",
]
typeCheckingMode = "strict"
pythonPlatform = "All"
pythonVersion = "3.13"
reportMissingImports = true
reportMissingTypeStubs = false
